+++ src/main/java/com/service/FilmServiceImpl.java
@@ -8,6 +8,7 @@ import com.domain.Filmshow;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Service
@@ -40,7 +41,7 @@ public class FilmServiceImpl implements FilmService {
     }
 
     @Override
-    public Film getFilmById(int id) throws DaoException {
+    public Film getFilmById(BigInteger id) throws DaoException {
         return filmDao.getFilmById(id);
     }
